Removed Controls from CorvEngine and instead re-did them because the old ones sucked. Started Main Menu as well as Interface. Still a lot to do involving these, so will leave the rest to Brown.
To view the Main Menu, go to CorvGame -> SceneManager and uncomment the one line that pushes the state to MainMenu. Don't forget to comment out the PushState(SceneManager) line.

Also, i'm not exactly sure if this is the best way to do the menu and such. For now, should be fine until Onion says something. :P